* master:
  Don't rely on SSLNetVC when HttpSM gathers info about SSL (#7961)
  conf_remap: demote 'Invalid configuration' to warning (#7991)
  Cleans up the code bit, including milliseconds consistency (#7989)
  Pass through expect header and handle 100-continue response (#7962)
  Treat TRACE with body as bad request (#7905)
  Thread safe Mersenne Twister 64 using c++11 (#7859)
  ESI plugin documentation updates. (#7970)
  Add log name configuration and stderr/stdout support. (#7937)
  Cleanup: Constify MIMEHdr (#7949)
  Fixed compile error with Linux AIO unit test (#7958)
  Note YAML parser library bug, and work-around, in documentation. (#7963)
  Ensure that the content-length value is only digits (#7964)
  String the url fragment for outgoing requests (#7966)
  Fix for HTTP/2 frames (#7965)
  Improve parsing error messages for strategies.yaml. (#7948)
  fix the scheme of h2 0rtt tests (#7957)
  Fix double test flakiness due to EOS/TXN_CLOSE race (#7956)
  Use proxy.config.log.hostname for rotated log filenames (#7943)
  Fixed memory leak in the QUIC stream manager (#7951)
  Fixup TS_USE_LINUX_NATIVE_AIO AIO_MODE_NATIVE (#7832)
  Update GitHub stale action to auto close old PRs (#7952)
  Revert "Do not invalidate cached resources upon error responses to unsafe methods (#7864)" (#7954)
  regex_revalidate: add stats for miss/stale counts (#7950)
  Do not invalidate cached resources upon error responses to unsafe methods (#7864)
  Add an HTTP/2 304 "Not Modified" AuTest. (#7882)
  regex_revalidate: optionally retain rule epoch state across restarts (#7939)
  Fixed memory leak in QUIC ack frame unit test (#7947)
  cache_promote: Don't promote on uncacheable requests (#7942)
  Fix dynamic-stack-buffer-overflow of cachekey plugin (#7945)
  Compilation error fixes for QUIC unit tests (#7944)
  Adds bytes counting as a trigger to the cache_promote LRU (#7765)
  Add a JSON schema for strategies.yaml (#7932)
  Remove second call to TRANSACT_RETURN while handling cache write lock (#7873)
  Close connection after every bad request for HTTP/1.1 (#7885)
  Pin Sphinx to 3.x to unblock `make html` (#7940)
  Add support for Remap rule hit stats (#7936)
  Remove scrap log object dead code (#7935)
  Add STL forward iterators to DLL container. (#7934)
  Add log SQUID code testing to redirect.test.py Au test. (#7870)
  Fix race condition on server session state (#7921)
  regex_reval: bug where rule type is always reported as the first (#7928)
  Remove duplicate entry in overridable txn vars. (#7930)
  Satisfy ci/jenkins/bin/clang-format.sh (#7929)
  Add a basic Au test using strategies.yaml, with consistent hashing. (#7911)
  Add a chunked negative revalidating test. (#7907)
  Ensure that URL components are valid when alternate eviction is logged (#7924)
  fix grammar (#7927)
  AuTest: Enable h2spec generic test cases (#7926)
  Adjust vc read errors (#7923)
  Remove bucket search from IntrusiveHashMap::erase (#7848)
  Ensure TS_VCONN_CLOSE_HOOK hook is called during TS_EVENT_VCONN_CLOSE. (#7913)
  Update docs languages file to add 9.1.x for en and ja (#7917)
  * Adds a new peering ring mode to next hop selection strategies. (#7897)
  Add Au test for strategies.yaml, with consistent hashing, with fallover. (#7914)
  Make HttpSM server reference a Transaction instead of a Session (#7849)
  Set accept_options of Http1Transaction in Http1ClientSession::new_connection() (#7894)
  Reset Http1Transaction before adding vc to keep_alive_queue (#7892)
  Add dead server policy control and metric. Improve messages. (#7757)
  Ensure the HTTP protion of the protocol string is upper case (#7904)
  Fixed spelling mistakes in the docs (#7896)
  add MISS capability to the regex_revalidate plugin (#7899)
  docs: fix capitalization of Linux (#7898)
  Redirect - Make TS to honour the number_of_redirections configuration value (#7867)
  Clean up producer more regularly (#7386)
  Fix crash in open_close_h2 (#7586)
  Cleanup Http2ClientSession SessionHandler (#7876)
  Enforce HTTP parsing restrictions on HTTP versions supported (#7875)
  Do not delete the continuation twice (#7862)
  Cleanup: refer Http2ClientSession::mutex (#7853)
  Autest - Proxy Verifier Extension, add context template $-base string substitution in the replay file. (#7866)
  Fixed some spelling mistakes in comments (#7869)
  Fixed ASAN issues with MMH test (#7868)
  Cleanup: Move member functions defined inside of class definitions of Http2ConnectionState & Http2ConnectionSettings (#7854)
  Add URI Signing cdnistd Claim Implementation (#7822)
  Adds a new --enable-all-asserts configure option (#7858)
  Unifdef test code for MMH and moved it into its own test file (#7841)
  Clean up lua plugin doc for overridable configurations (#7844)
  Save and propagate epoll network error (#7809)
  Add method to write an IpAddr value to a sockaddr. (#7821)
  Add proxy.config.http.max_proxy_cycles (#7657)
  Update NextHop strategies so that unavailable server retry codes (#7837)
  generator: allow for POST requests (#7635)
  Fixed double declaration types for log buffer tracking (#7847)
  Extra braces for clang 5 / ubuntu 16.04 on array initialization (#7842)

 Conflicts:
	iocore/net/quic/QUICStreamFactory.cc

+++ iocore/net/quic/test/test_QUICPacket.cc
@@ -249,7 +249,8 @@ TEST_CASE("Sending Packet", "[quic]")
     CHECK(packet.is_crypto_packet());
 
     packet.store(buf, &len);
-    CHECK(len == packet.size() - 16);
+    // See above packet.size() CHECK: packet.size must be greater than 16.
+    CHECK(len == static_cast<size_t>(packet.size() - 16));
     CHECK(memcmp(buf, expected, len - 16) == 0);
   }
